# Implement #3129 (Optimized history traversals in the REST API)