--- a/beacon_chain/beacon_node.nim
+++ b/beacon_chain/beacon_node.nim
@@ -23,7 +23,8 @@ import
     sync_committee_msg_pool],
   ./spec/datatypes/base,
   ./sync/[sync_manager, request_manager],
-  ./validators/[action_tracker, validator_monitor, validator_pool]
+  ./validators/[action_tracker, validator_monitor, validator_pool],
+  ./rpc/state_ttl_cache
 
 export
   osproc, chronos, httpserver, presto, action_tracker, beacon_clock,
@@ -70,6 +71,7 @@ type
     onAttestationSent*: OnAttestationCallback
     restKeysCache*: Table[ValidatorPubKey, ValidatorIndex]
     validatorMonitor*: ref ValidatorMonitor
+    stateTtlCache*: StateTtlCache
 
 const
   MaxEmptySlotCount* = uint64(10*60) div SECONDS_PER_SLOT