--- a/beacon_chain/rpc/rest_debug_api.nim
+++ b/beacon_chain/rpc/rest_debug_api.nim
@@ -9,7 +9,7 @@ import std/sequtils
 import chronicles
 import ".."/[version, beacon_node],
        ".."/spec/forks,
-       "."/rest_utils
+       "."/[rest_utils, state_ttl_cache]
 
 export rest_utils