--- a/benchmarks/rest_api_benchmark.nim.cfg
+++ b/benchmarks/rest_api_benchmark.nim.cfg
@@ -1,0 +1,2 @@
+-d:"chronicles_runtime_filtering=on"
+-d:"chronicles_disable_thread_id"