… docs/the_nimbus_book/src/options.md