Up to 10x speed-up for historic traversals through the API:

```
# Before
➜  nimbus-eth2 git:(feature/rest_api_optimized_history_traversal) build/rest_api_benchmark --start-slot=9000 --count=250
INF 2021-12-22 09:24:55.962+02:00 Benchmarking ...                           file=rest_api_benchmark.nim:42 apiName=getStateRoot
INF 2021-12-22 09:26:08.637+02:00 getStateRoot                               file=rest_api_benchmark.nim:43 t=72.67443995799999s
INF 2021-12-22 09:26:08.637+02:00 Benchmarking ...                           file=rest_api_benchmark.nim:42 apiName=getStateFork
INF 2021-12-22 09:27:20.277+02:00 getStateFork                               file=rest_api_benchmark.nim:43 t=71.639665333s
INF 2021-12-22 09:27:20.277+02:00 Benchmarking ...                           file=rest_api_benchmark.nim:42 apiName=getStateFinalityCheckpoints
INF 2021-12-22 09:28:33.398+02:00 getStateFinalityCheckpoints                file=rest_api_benchmark.nim:43 t=73.12129654100001s
INF 2021-12-22 09:28:33.398+02:00 Benchmarking ...                           file=rest_api_benchmark.nim:42 apiName=getStateValidatorBalances
INF 2021-12-22 09:29:55.869+02:00 getStateValidatorBalances                  file=rest_api_benchmark.nim:43 t=82.470523375s

# After
➜  nimbus-eth2 git:(feature/rest_api_optimized_history_traversal) build/rest_api_benchmark --start-slot=9000 --count=250
INF 2021-12-22 09:24:55.962+02:00 Benchmarking ...                           file=rest_api_benchmark.nim:42 apiName=getStateRoot
INF 2021-12-22 09:26:08.637+02:00 getStateRoot                               file=rest_api_benchmark.nim:43 t=6.67443995799999s
INF 2021-12-22 09:21:32.563+02:00 Benchmarking ...                           file=rest_api_benchmark.nim:42 apiName=getStateFork
INF 2021-12-22 09:21:39.626+02:00 getStateFork                               file=rest_api_benchmark.nim:43 t=7.062567791s
INF 2021-12-22 09:21:39.626+02:00 Benchmarking ...                           file=rest_api_benchmark.nim:42 apiName=getStateFinalityCheckpoints
INF 2021-12-22 09:21:46.180+02:00 getStateFinalityCheckpoints                file=rest_api_benchmark.nim:43 t=6.55376425s
INF 2021-12-22 09:21:46.180+02:00 Benchmarking ...                           file=rest_api_benchmark.nim:42 apiName=getStateValidatorBalances
INF 2021-12-22 09:22:01.645+02:00 getStateValidatorBalances                  file=rest_api_benchmark.nim:43 t=15.465585083s
```