@arnetheduck 2021-12-22
I've keep wondering why `Implement #3129` as the main headline for a git commit makes sense - in nim-lang, this is also the culture, what's the .. idea? 

The log message is written _for_ the consumer of it, and when reading the history (`git log --oneline`), 3129 isn't meaningful / helpful - it seems to me that `Bla bla (implements #XXXX)` is more useful to readers?