@arnetheduck 2021-12-28
fwiw, most comments on the PR are of the "value-on-the-table" type, except the loop condition thing - with that fixed, we can merge it (but it would be a shame, the comments will end up forgotten)